requirements.txt
Pillow
torch
torchvision
numpy
joblib
